HOCS-4148: simplify Home Secretary tagging logic
This change simplifies the Home Secretary tagging logic to just whether
the `HomeSecReply` field is present. Previously, this was
over-complicated and relied on differing stages.
