+++ src/main/java/uk/gov/digital/ho/hocs/casework/api/StageService.java
@@ -24,7 +24,6 @@ import uk.gov.digital.ho.hocs.casework.domain.model.ActiveStage;
 import uk.gov.digital.ho.hocs.casework.domain.model.BaseStage;
 import uk.gov.digital.ho.hocs.casework.domain.model.CaseData;
 import uk.gov.digital.ho.hocs.casework.domain.model.CaseDataTag;
-import uk.gov.digital.ho.hocs.casework.domain.model.SomuItem;
 import uk.gov.digital.ho.hocs.casework.domain.model.Stage;
 import uk.gov.digital.ho.hocs.casework.domain.model.StageWithCaseData;
 import uk.gov.digital.ho.hocs.casework.domain.repository.CaseTagRepository;
@@ -576,7 +575,7 @@ public class StageService {
             daysElapsedCalculator.updateDaysElapsed(stage.getData(), stage.getCaseDataType());
 
             //TODO: HOCS-5871 Remove after Workflow Implementation released
-            stage.appendTags(stageTagsDecorator.decorateTags(stage.getData(), stage.getStageType()));
+            stage.appendTags(stageTagsDecorator.decorateTags(stage.getData()));
         }
     }
